Split off session state handler and responder
