+++ app/Libraries/Session/Store.php
@@ -21,7 +21,7 @@
 namespace App\Libraries\Session;
 
 use App\Events\UserLogoutEvent;
-use App\Libraries\UserVerification;
+use App\Libraries\UserVerificationState;
 use Auth;
 use Illuminate\Support\Str;
 use Jenssegers\Agent\Agent;
@@ -194,7 +194,7 @@ class Store extends \Illuminate\Session\Store
             $sessionMeta[$id]['device'] = $agent->device();
             $sessionMeta[$id]['platform'] = $agent->platform();
             $sessionMeta[$id]['browser'] = $agent->browser();
-            $sessionMeta[$id]['verified'] = isset($session['verified']) && $session['verified'] === UserVerification::VERIFIED;
+            $sessionMeta[$id]['verified'] = isset($session['verified']) && $session['verified'] === UserVerificationState::VERIFIED;
         }
 
         // returns sessions sorted from most to least recently active
