+++ tests/Controllers/AccountControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Libraries\UserVerification;
+use App\Libraries\UserVerificationState;
 use App\Models\User;
 use App\Models\UserProfileCustomization;
 use App\Models\WeakPassword;
@@ -65,7 +65,7 @@ class AccountControllerTest extends TestCase
         $newEmail = 'new-'.$this->user->user_email;
 
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.email'), [
                 'user' => [
                     'current_password' => 'password',
@@ -85,7 +85,7 @@ class AccountControllerTest extends TestCase
         $newEmail = 'new-'.$this->user->user_email;
 
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.email'), [
                 'user' => [
                     'current_password' => 'password1',
@@ -101,7 +101,7 @@ class AccountControllerTest extends TestCase
         $newPassword = 'newpassword';
 
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'password',
@@ -117,7 +117,7 @@ class AccountControllerTest extends TestCase
     public function testUpdatePasswordInvalidCurrentPassword()
     {
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'notpassword',
@@ -131,7 +131,7 @@ class AccountControllerTest extends TestCase
     public function testUpdatePasswordInvalidPasswordConfirmation()
     {
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'password',
@@ -145,7 +145,7 @@ class AccountControllerTest extends TestCase
     public function testUpdatePasswordUsernameAsPassword()
     {
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'password',
@@ -159,7 +159,7 @@ class AccountControllerTest extends TestCase
     public function testUpdatePasswordShortPassword()
     {
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'password',
@@ -177,7 +177,7 @@ class AccountControllerTest extends TestCase
         WeakPassword::add($weakPassword);
 
         $this->actingAs($this->user())
-            ->withSession(['verified' => UserVerification::VERIFIED])
+            ->withSession(['verified' => UserVerificationState::VERIFIED])
             ->json('PUT', route('account.password'), [
                 'user' => [
                     'current_password' => 'password',
